+++ ql/src/java/org/apache/hadoop/hive/ql/parse/TaskCompilerFactory.java
@@ -46,7 +46,7 @@ public class TaskCompilerFactory {
       case IMPALA:
         return new ImpalaCompiler(
                 conf.getVar(HiveConf.ConfVars.HIVE_IMPALA_EXECUTION_MODE).equals("plan"),
-                conf.getIntVar(HiveConf.ConfVars.HIVE_IMPALA_FETCH_SIZE));
+                conf.getLongVar(HiveConf.ConfVars.HIVE_IMPALA_FETCH_SIZE));
       case INVALID_ENGINE:
         throw new UnsupportedOperationException("Invalid execution engine specified.");
     }
